// apps/studio/src/protocol/types.ts
export type CamId = 'cam1' | 'cam2' | 'cam3' | 'cam4';
export type TargetId = CamId | 'p4';

export const CAM_IDS: CamId[] = ['cam1', 'cam2', 'cam3', 'cam4'];
// ...
export type FlashLevel = 'low' | 'medium' | 'high';
export type FlashDistance = '0.5-1' | '1-2' | '2-3';